requirements.txt
arrow==0.16.0
attrs==20.1.0
aws-lambda-builders==1.0.0
aws-sam-cli==1.1.0
aws-sam-translator==1.26.0
awscli==1.18.124
binaryornot==0.4.4
boto3==1.14.47
botocore==1.17.47
certifi==2020.6.20
chardet==3.0.4
chevron==0.13.1
click==7.1.2
colorama==0.4.3
cookiecutter==1.6.0
dateparser==0.7.6
docker==4.2.2
docutils==0.15.2
Flask==1.0.4
flask-lambda-python36==0.1.0
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
poyo==0.5.0
pyasn1==0.4.8
pypiwin32==223
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
pywin32==228
PyYAML==5.3.1
regex==2020.7.14
requests==2.23.0
rsa==4.5
s3transfer==0.3.3
serverlessrepo==0.1.9
six==1.15.0
tomlkit==0.5.8
tzlocal==2.1
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
zipp==3.1.0
